guitarEffects/requirements.txt
Django>=4.0
psycopg2-binary
openai==1.95.1
python-dotenv
faster-whisper